refactor:커스텀 예외 폴더 이름 변경 및 폴더구조 변경 -#13
