+++ src/main/java/com/mju_lion/letter/exception/ConflictException.java
@@ -1,6 +1,6 @@
-package com.mju_lion.letter.error.exception;
+package com.mju_lion.letter.exception;
 
-import com.mju_lion.letter.error.ErrorCode;
+import com.mju_lion.letter.exception.errorcode.ErrorCode;
 
 public class ConflictException extends CustomException {
     public ConflictException(ErrorCode errorCode){
