+++ src/main/java/com/mju_lion/letter/exception/CustomException.java
@@ -1,7 +1,7 @@
-package com.mju_lion.letter.error.exception;
+package com.mju_lion.letter.exception;
 
 
-import com.mju_lion.letter.error.ErrorCode;
+import com.mju_lion.letter.exception.errorcode.ErrorCode;
 import lombok.Getter;
 
 @Getter
