+++ src/main/java/com/mju_lion/letter/exception/NotFoundException.java
@@ -1,6 +1,6 @@
-package com.mju_lion.letter.error.exception;
+package com.mju_lion.letter.exception;
 
-import com.mju_lion.letter.error.ErrorCode;
+import com.mju_lion.letter.exception.errorcode.ErrorCode;
 
 public class NotFoundException extends CustomException {
     public NotFoundException(ErrorCode errorCode) {
