+++ src/main/java/com/mju_lion/letter/exception/UnauthorizedException.java
@@ -1,6 +1,6 @@
-package com.mju_lion.letter.error.exception;
+package com.mju_lion.letter.exception;
 
-import com.mju_lion.letter.error.ErrorCode;
+import com.mju_lion.letter.exception.errorcode.ErrorCode;
 
 public class UnauthorizedException extends CustomException {
     public UnauthorizedException(ErrorCode errorCode) {
