+++ src/main/java/com/mju_lion/letter/service/AuthService.java
@@ -4,10 +4,10 @@ import com.mju_lion.letter.authentication.PasswordHashEncryption;
 import com.mju_lion.letter.dto.request.auth.LoginDto;
 import com.mju_lion.letter.dto.request.auth.SignupDto;
 import com.mju_lion.letter.entity.User;
-import com.mju_lion.letter.error.ErrorCode;
-import com.mju_lion.letter.error.exception.ConflictException;
-import com.mju_lion.letter.error.exception.NotFoundException;
-import com.mju_lion.letter.error.exception.UnauthorizedException;
+import com.mju_lion.letter.exception.errorcode.ErrorCode;
+import com.mju_lion.letter.exception.ConflictException;
+import com.mju_lion.letter.exception.NotFoundException;
+import com.mju_lion.letter.exception.UnauthorizedException;
 import com.mju_lion.letter.repository.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
