+++ src/main/java/com/mju_lion/letter/service/BoardService.java
@@ -6,9 +6,9 @@ import com.mju_lion.letter.dto.response.board.BoardResponseData;
 import com.mju_lion.letter.entity.Board;
 import com.mju_lion.letter.entity.Comment;
 import com.mju_lion.letter.entity.User;
-import com.mju_lion.letter.error.ErrorCode;
-import com.mju_lion.letter.error.exception.ForbiddenException;
-import com.mju_lion.letter.error.exception.NotFoundException;
+import com.mju_lion.letter.exception.errorcode.ErrorCode;
+import com.mju_lion.letter.exception.ForbiddenException;
+import com.mju_lion.letter.exception.NotFoundException;
 import com.mju_lion.letter.repository.BoardRepository;
 import com.mju_lion.letter.repository.CommentRepository;
 import lombok.RequiredArgsConstructor;
@@ -19,7 +19,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
