+++ src/main/java/com/mju_lion/letter/service/CommentService.java
@@ -6,9 +6,9 @@ import com.mju_lion.letter.dto.response.comment.CommentResponseData;
 import com.mju_lion.letter.entity.Board;
 import com.mju_lion.letter.entity.Comment;
 import com.mju_lion.letter.entity.User;
-import com.mju_lion.letter.error.ErrorCode;
-import com.mju_lion.letter.error.exception.ForbiddenException;
-import com.mju_lion.letter.error.exception.NotFoundException;
+import com.mju_lion.letter.exception.errorcode.ErrorCode;
+import com.mju_lion.letter.exception.ForbiddenException;
+import com.mju_lion.letter.exception.NotFoundException;
 import com.mju_lion.letter.repository.BoardRepository;
 import com.mju_lion.letter.repository.CommentRepository;
 import lombok.RequiredArgsConstructor;
